fix(refactor): delete now uses flux
